requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
idna==3.4
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.0
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
